Move features to proper package
