+++ app/src/main/java/com/example/appName/presentation/features/dummy/DummyView.kt
@@ -1,4 +1,4 @@
-package com.example.appName.presentation.dummy
+package com.example.appName.presentation.features.dummy
 
 import io.reactivex.Observable
